requirements-dev.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.0.5
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-pure-pagination==0.3.0
httplib2==0.11.3
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2==1.9.0.post1
oauthlib==2.1.0
pycryptodome==3.6.1
Pygments==2.2.0
PyJWT==1.6.4
pylint==1.9.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
urllib3==1.22
wrapt==1.10.11
